Split api tests in two files
